ZOOKEEPER-1841. problem in QuorumTest (Germán via fpj)
git-svn-id: https://svn.apache.org/repos/asf/zookeeper/branches/branch-3.4@1551930 13f79535-47bb-0310-9956-ffa450edef68
(cherry picked from commit 54eefdb4fc57e18f28bbb520e4abb163de2f936a)

Author: Flavio Paiva Junqueira <fpj@apache.org>
Reason: Bug
Ref: CDH-63687
