+++ CHANGES.txt
@@ -168,6 +168,8 @@ IMPROVEMENTS:
   ZOOKEEPER-2360: Update commons collections version used by tests/releaseaudit
   (phunt via cnauroth)
 
+  ZOOKEEPER-1841. problem in QuorumTest (Germán via fpj)
+
 IMPROVEMENTS:
 
   ZOOKEEPER-1583. Document maxClientCnxns in conf/zoo_sample.cfg
